[Cart] Simplify cart recalculation after customer cart assignment
