+++ src/Handler/Cart/AssignCustomerToCartHandler.php
@@ -6,6 +6,7 @@ namespace Sylius\ShopApiPlugin\Handler\Cart;
 
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Repository\OrderRepositoryInterface;
+use Sylius\Component\Order\Processor\OrderProcessorInterface;
 use Sylius\ShopApiPlugin\Command\Cart\AssignCustomerToCart;
 use Sylius\ShopApiPlugin\Provider\CustomerProviderInterface;
 use Webmozart\Assert\Assert;
@@ -13,26 +14,35 @@ use Webmozart\Assert\Assert;
 final class AssignCustomerToCartHandler
 {
     /** @var OrderRepositoryInterface */
-    private $orderRepository;
+    private $cartRepository;
+
+    /** @var OrderProcessorInterface */
+    private $orderProcessor;
 
     /** @var CustomerProviderInterface */
     private $customerProvider;
 
-    public function __construct(OrderRepositoryInterface $orderRepository, CustomerProviderInterface $customerProvider)
-    {
-        $this->orderRepository = $orderRepository;
+    public function __construct(
+        OrderRepositoryInterface $cartRepository,
+        OrderProcessorInterface $orderProcessor,
+        CustomerProviderInterface $customerProvider
+    ) {
+        $this->cartRepository = $cartRepository;
         $this->customerProvider = $customerProvider;
+        $this->orderProcessor = $orderProcessor;
     }
 
     public function __invoke(AssignCustomerToCart $assignOrderToCustomer): void
     {
-        /** @var OrderInterface $order */
-        $order = $this->orderRepository->findOneBy(['tokenValue' => $assignOrderToCustomer->orderToken()]);
+        /** @var OrderInterface $cart */
+        $cart = $this->cartRepository->findOneBy(['tokenValue' => $assignOrderToCustomer->orderToken()]);
 
-        Assert::notNull($order, sprintf('Order with %s token has not been found.', $assignOrderToCustomer->orderToken()));
+        Assert::notNull($cart, sprintf('Order with %s token has not been found.', $assignOrderToCustomer->orderToken()));
 
         $customer = $this->customerProvider->provide($assignOrderToCustomer->email());
 
-        $order->setCustomer($customer);
+        $cart->setCustomer($customer);
+
+        $this->orderProcessor->process($cart);
     }
 }
