+++ src/Resources/config/services/handler/cart.xml
@@ -101,6 +101,7 @@
         <service id="sylius.shop_api_plugin.handler.assign_customer_to_cart_handler"
                  class="Sylius\ShopApiPlugin\Handler\Cart\AssignCustomerToCartHandler">
             <argument type="service" id="sylius.repository.order"/>
+            <argument type="service" id="sylius.order_processing.order_processor" />
             <argument type="service" id="sylius.shop_api_plugin.provider.customer_provider" />
             <tag name="messenger.message_handler" />
         </service>
